requirements.txt
requests==2.18.4
numpy==1.14.5
Keras==2.2.1
PyYAML==3.13
